refactor(tools-expose-variables): overhaul
